+++ game/controllers/player_controller.h
@@ -5,8 +5,9 @@
 #pragma once
 
 #include "config.h"
+#include "state_machines_list.h"
 
-typedef void (*actor_move_func)(Verse::Config&, Verse::EntityID);
+typedef void (*actor_move_func)(Verse::Config&, Verse::EntityID, Verse::State::StateType);
 
 namespace Verse::Controller::Player
 {